Updates to M1 lesson 2
For this lesson since there are so many screenshots, I've created new ones and named them the same as the originals. I just added "v2" at the end of the file names. There was one new screenshot added and I put a note as to where it should be inserted. File name: editor-add-new-template-popup.png